backend/requirements.txt
flask
flask-cors
langchain
pypdf
faiss-cpu
openai
python-dotenv
PyPDF2